btlp: the cast, the stage close, an enemy's artwork, and three spell starts
BtlMemberMotion06, BtlStageClose and BtlLoadEnemyGfx come out of their
guards, as do BtlFxStart38, BtlFxStart88 and BtlFxStart8C. memberact's
.rodata segment now starts at BtlMemberMotion06's jump table.
ovl_btlp_entry goes from 99.04% to 99.29% and BtlBeginTalking from 99.70%
to 99.94%, both still guarded.

Three levers are written up in docs/Matching Nudges.md: a range test
written on the field itself keeps the load's copy and grows the frame; a
constant in a branch's delay slot can come from the fall-through; and a
call that ends its basic block keeps its arguments in expand order.
docs/RTL Dumps.md covers asking cc1 which of its passes made a decision.

+++ src/btlp/pickable.c
@@ -68,7 +68,7 @@ int BtlAnyMemberTargetable(void)
 {
     /* Eight bytes of locals the routine reserves and never writes. */
     long unused[2];
-    short status;
+    int  status;
     int  i;
 
     i = 0;
